refactor player interaction codes
