+++ model/game.py
@@ -7,3 +7,8 @@ class Game:
 
     def show(self, my_plt):
         pass
+
+    def next_turn_and_show(self, i, my_plt):
+        self.next_turn()
+        print(i, "turn")
+        self.show(my_plt)
