+++ model/plotter.py
@@ -1,3 +1,4 @@
+# unused for now
 class Plotter:
     def plot_by_input(self, location, color):
         pass
